Move literals to //fly/types
