+++ fly/socket/socket_config.cpp
@@ -1,6 +1,6 @@
 #include "fly/socket/socket_config.h"
 
-#include "fly/literals.h"
+#include "fly/types/literals.h"
 
 namespace fly {
