+++ fly/types/bit_stream/bit_stream_writer.cpp
@@ -1,7 +1,7 @@
 #include "fly/types/bit_stream/bit_stream_writer.h"
 
-#include "fly/literals.h"
 #include "fly/types/bit_stream/detail/bit_stream_constants.h"
+#include "fly/types/literals.h"
 
 namespace fly {
