+++ test/coders/huffman.cpp
@@ -1,8 +1,8 @@
 #include "fly/coders/huffman/huffman_config.h"
 #include "fly/coders/huffman/huffman_decoder.h"
 #include "fly/coders/huffman/huffman_encoder.h"
-#include "fly/literals.h"
 #include "fly/types/bit_stream/bit_stream_writer.h"
+#include "fly/types/literals.h"
 #include "fly/types/string/string.h"
 #include "test/util/path_util.h"
