+++ test/config/config_manager.cpp
@@ -1,9 +1,9 @@
 #include "fly/config/config_manager.h"
 
 #include "fly/config/config.h"
-#include "fly/literals.h"
 #include "fly/path/path_config.h"
 #include "fly/task/task_manager.h"
+#include "fly/types/literals.h"
 #include "fly/types/string/string.h"
 #include "test/config/test_config.h"
 #include "test/util/path_util.h"
