+++ test/types/bit_stream.cpp
@@ -1,7 +1,7 @@
-#include "fly/literals.h"
 #include "fly/types/bit_stream/bit_stream_reader.h"
 #include "fly/types/bit_stream/bit_stream_writer.h"
 #include "fly/types/bit_stream/detail/bit_stream_constants.h"
+#include "fly/types/literals.h"
 
 #include <gtest/gtest.h>
